Moved test-module to its own directory
